@@ -4,29 +4,28 @@ using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Concurrent;
 
-namespace MonoLibrary.Dependency.Loggers.ConsoleLogger
+namespace MonoLibrary.Dependency.Loggers.ConsoleLogger;
+
+[ProviderAlias("ColorConsole")]
+public sealed class ColorConsoleLoggerProvider : ILoggerProvider
 {
-    [ProviderAlias("ColorConsole")]
-    public sealed class ColorConsoleLoggerProvider : ILoggerProvider
+    private readonly IDisposable _onChangeToken;
+    private ColorConsoleLoggerConfiguration _currentConfig;
+    private readonly ConcurrentDictionary<string, ColorConsoleLogger> _loggers = new(StringComparer.OrdinalIgnoreCase);
+
+    public ColorConsoleLoggerProvider(IOptionsMonitor<ColorConsoleLoggerConfiguration> config)
     {
-        private readonly IDisposable _onChangeToken;
-        private ColorConsoleLoggerConfiguration _currentConfig;
-        private readonly ConcurrentDictionary<string, ColorConsoleLogger> _loggers = new(StringComparer.OrdinalIgnoreCase);
+        _currentConfig = config.CurrentValue;
+        _onChangeToken = config.OnChange(updatedConfig => _currentConfig = updatedConfig);
+    }
 
-        public ColorConsoleLoggerProvider(IOptionsMonitor<ColorConsoleLoggerConfiguration> config)
-        {
-            _currentConfig = config.CurrentValue;
-            _onChangeToken = config.OnChange(updatedConfig => _currentConfig = updatedConfig);
-        }
+    public ILogger CreateLogger(string categoryName) => _loggers.GetOrAdd(categoryName, name => new ColorConsoleLogger(name, GetCurrentConfig));
 
-        public ILogger CreateLogger(string categoryName) => _loggers.GetOrAdd(categoryName, name => new ColorConsoleLogger(name, GetCurrentConfig));
+    private ColorConsoleLoggerConfiguration GetCurrentConfig() => _currentConfig;
 
-        private ColorConsoleLoggerConfiguration GetCurrentConfig() => _currentConfig;
-
-        public void Dispose()
-        {
-            _loggers.Clear();
-            _onChangeToken?.Dispose();
-        }
+    public void Dispose()
+    {
+        _loggers.Clear();
+        _onChangeToken?.Dispose();
     }
 }
