@@ -1,15 +1,9 @@
 ﻿using System;
 
-namespace MonoLibrary.Dependency
-{
-    [AttributeUsage(AttributeTargets.Class, Inherited = true, AllowMultiple = false)]
-    public sealed class SettingsAttribute : Attribute
-    {
-        public string Name { get; }
+namespace MonoLibrary.Dependency;
 
-        public SettingsAttribute(string name = null)
-        {
-            Name = name;
-        }
-    }
+[AttributeUsage(AttributeTargets.Class, Inherited = true, AllowMultiple = false)]
+public sealed class SettingsAttribute(string name = null) : Attribute
+{
+    public string Name { get; } = name;
 }
