@@ -3,34 +3,28 @@
 using MonoLibrary.Engine.Components.Interfaces;
 using MonoLibrary.Engine.Objects;
 
-namespace MonoLibrary.Engine.Components.Colliders
+namespace MonoLibrary.Engine.Components.Colliders;
+
+public class CircleColliderComponent(GameObject owner) : ColliderComponent(owner)
 {
-    public class CircleColliderComponent : ColliderComponent
+    public override Rectangle Bounds => new((Offset + Owner.Position - new Vector2(Radius)).ToPoint(), new Point((int)Radius * 2));
+    public Vector2 Offset { get; set; }
+    public float Radius { get; set; }
+
+    public override bool IsColliding(IColliderComponent other)
     {
-        public override Rectangle Bounds => new((Offset + Owner.Position - new Vector2(Radius)).ToPoint(), new Point((int)Radius * 2));
-        public Vector2 Offset { get; set; }
-        public float Radius { get; set; }
+        if (other is CircleColliderComponent circle)
+            return ColliderResolver.IsColliding(this, circle);
+        else if (other is AABBColliderComponent aabb)
+            return ColliderResolver.IsColliding(aabb, this);
+        else if (other is CompositeColliderComponent composite)
+            return composite.IsColliding(this);
 
-        public CircleColliderComponent(GameObject owner) : base(owner)
-        {
+        return false;
+    }
 
-        }
-
-        public override bool IsColliding(IColliderComponent other)
-        {
-            if (other is CircleColliderComponent circle)
-                return ColliderResolver.IsColliding(this, circle);
-            else if (other is AABBColliderComponent aabb)
-                return ColliderResolver.IsColliding(aabb, this);
-            else if (other is CompositeColliderComponent composite)
-                return composite.IsColliding(this);
-
-            return false;
-        }
-
-        public override ColliderRenderer GetRenderer()
-        {
-            return new CircleColliderRenderer(Owner, this);
-        }
+    public override ColliderRenderer GetRenderer()
+    {
+        return new CircleColliderRenderer(Owner, this);
     }
 }
