@@ -1,11 +1,10 @@
 ﻿using MonoLibrary.Engine.Objects;
 
-namespace MonoLibrary.Engine.Components.Interfaces
-{
-    public interface IComponent
-    {
-        GameObject Owner { get; }
+namespace MonoLibrary.Engine.Components.Interfaces;
 
-        void OnDestroy() { }
-    }
+public interface IComponent
+{
+    GameObject Owner { get; }
+
+    void OnDestroy() { }
 }
