@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace MonoLibrary.Engine.Network.Components
-{
-    public interface IInvokable : IDisposable
-    {
-        void Invoke();
-    }
+namespace MonoLibrary.Engine.Network.Components;
 
-    public interface IInvokable<T> : IDisposable
-        where T : struct
-    {
-        void Invoke(T data);
-    }
+public interface IInvokable : IDisposable
+{
+    void Invoke();
+}
+
+public interface IInvokable<T> : IDisposable
+    where T : struct
+{
+    void Invoke(T data);
 }
