@@ -1,15 +1,14 @@
 ﻿using System.Net;
 
-namespace MonoLibrary.Engine.Network.Messages
+namespace MonoLibrary.Engine.Network.Messages;
+
+public class ServerDiscoveryRequest
 {
-    public class ServerDiscoveryRequest
-    {
 
-    }
+}
 
-    public class ServerDiscoveryResponse
-    {
-        public IPEndPoint ServerEndPoint { get; set; }
-        public string ConnectionKey { get; set; }
-    }
+public class ServerDiscoveryResponse
+{
+    public IPEndPoint ServerEndPoint { get; set; }
+    public string ConnectionKey { get; set; }
 }
