@@ -1,10 +1,9 @@
-﻿namespace MonoLibrary.Engine.Network.Messages
-{
-    public class SpawnRequest
-    {
-        public int PrefabId { get; set; }
-        public int NetworkOwnerId { get; set; }
+﻿namespace MonoLibrary.Engine.Network.Messages;
 
-        public byte[] Data { get; set; }
-    }
+public class SpawnRequest
+{
+    public int PrefabId { get; set; }
+    public int NetworkOwnerId { get; set; }
+
+    public byte[] Data { get; set; }
 }
