@@ -1,8 +1,7 @@
-﻿namespace MonoLibrary.Engine.Network.Messages
+﻿namespace MonoLibrary.Engine.Network.Messages;
+
+public class UpdateMessage
 {
-    public class UpdateMessage
-    {
-        public int[] NetIds { get; set; }
-        public byte[] Data { get; set; }
-    }
+    public int[] NetIds { get; set; }
+    public byte[] Data { get; set; }
 }
