@@ -9,48 +9,41 @@ using MonoLibrary.Engine.Objects;
 using System;
 using System.Collections.Generic;
 
-namespace MonoLibrary.Engine.Network
+namespace MonoLibrary.Engine.Network;
+
+public abstract class NetworkPrefabFactoryBase(IOptions<NetworkSettings> settings, ILogger logger) : INetworkFactory
 {
-    public abstract class NetworkPrefabFactoryBase : INetworkFactory
+    public delegate GameObject SpawnPrefab(NetDataReader reader);
+
+    protected readonly ILogger Logger = logger;
+    private readonly Dictionary<int, SpawnPrefab> _prefabs = [];
+
+    public bool IsServer { get; } = settings.Value.IsServer;
+
+    protected void RegisterPrefab(int prefabId, SpawnPrefab factory)
     {
-        public delegate GameObject SpawnPrefab(NetDataReader reader);
+        if (_prefabs.TryGetValue(prefabId, out var old))
+            Logger.LogWarning("Prefab with Id {id} already registered. Prefab '{new}' will override '{old}'.", prefabId, old.Method.Name, factory.Method.Name);
 
-        protected readonly ILogger Logger;
-        private readonly Dictionary<int, SpawnPrefab> _prefabs = new();
+        _prefabs[prefabId] = factory;
+        Logger.LogInformation("Prefab '{factory}' registered with Id {id}.", factory.Method.Name, prefabId);
+    }
 
-        public bool IsServer { get; }
+    /// <summary>
+    /// Helper to use <see cref="Enum"/> values.
+    /// </summary>
+    /// <typeparam name="TEnum"></typeparam>
+    /// <param name="prefabId"></param>
+    /// <param name="factory"></param>
+    protected void RegisterPrefab<TEnum>(TEnum prefabId, SpawnPrefab factory) where TEnum : struct, Enum
+    {
+        RegisterPrefab((int)Convert.ChangeType(prefabId, prefabId.GetTypeCode()), factory);
+    }
 
-        protected NetworkPrefabFactoryBase(IOptions<NetworkSettings> settings, ILogger logger)
-        {
-            IsServer = settings.Value.IsServer;
-            Logger = logger;
-        }
-
-        protected void RegisterPrefab(int prefabId, SpawnPrefab factory)
-        {
-            if (_prefabs.TryGetValue(prefabId, out var old))
-                Logger.LogWarning("Prefab with Id {id} already registered. Prefab '{new}' will override '{old}'.", prefabId, old.Method.Name, factory.Method.Name);
-
-            _prefabs[prefabId] = factory;
-            Logger.LogInformation("Prefab '{factory}' registered with Id {id}.", factory.Method.Name, prefabId);
-        }
-
-        /// <summary>
-        /// Helper to use <see cref="Enum"/> values.
-        /// </summary>
-        /// <typeparam name="TEnum"></typeparam>
-        /// <param name="prefabId"></param>
-        /// <param name="factory"></param>
-        protected void RegisterPrefab<TEnum>(TEnum prefabId, SpawnPrefab factory) where TEnum : struct, Enum
-        {
-            RegisterPrefab((int)Convert.ChangeType(prefabId, prefabId.GetTypeCode()), factory);
-        }
-
-        public GameObject Spawn(int prefabId, NetDataReader reader)
-        {
-            SpawnPrefab prefab = _prefabs[prefabId];
-            Logger.LogTrace("Spawning prefab {prefabId} '{prefabName}', with {bytes} bytes of data.", prefabId, prefab.Method.Name, reader?.RawDataSize ?? 0);
-            return prefab.Invoke(reader);
-        }
+    public GameObject Spawn(int prefabId, NetDataReader reader)
+    {
+        SpawnPrefab prefab = _prefabs[prefabId];
+        Logger.LogTrace("Spawning prefab {prefabId} '{prefabName}', with {bytes} bytes of data.", prefabId, prefab.Method.Name, reader?.RawDataSize ?? 0);
+        return prefab.Invoke(reader);
     }
 }
