@@ -1,23 +1,17 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 
-namespace MonoLibrary.Engine.Pools
+namespace MonoLibrary.Engine.Pools;
+
+[SkipLocalsInit]
+public readonly struct Pooled<T>(Action<T> returnPool, T item) : IPooled<T>
 {
-    [SkipLocalsInit]
-    public readonly struct Pooled<T> : IPooled<T>
+    private readonly Action<T> _return = returnPool;
+
+    public T Item { get; } = item;
+
+    public void Dispose()
     {
-        private readonly Action<T> _return;
-        public T Item { get; }
-
-        public Pooled(Action<T> returnPool, T item)
-        {
-            _return = returnPool;
-            Item = item;
-        }
-
-        public void Dispose()
-        {
-            _return.Invoke(Item);
-        }
+        _return.Invoke(Item);
     }
 }
