@@ -1,21 +1,14 @@
 ﻿using System;
 
-namespace MonoLibrary.Engine.Services.Helpers
+namespace MonoLibrary.Engine.Services.Helpers;
+
+public readonly struct Subscription<T>(T item, Action<T> unsubscribe) : IDisposable
 {
-    public readonly struct Subscription<T> : IDisposable
+    private readonly T _item = item;
+    private readonly Action<T> _unsubscribe = unsubscribe;
+
+    public void Dispose()
     {
-        private readonly T _item;
-        private readonly Action<T> _unsubscribe;
-
-        public Subscription(T item, Action<T> unsubscribe)
-        {
-            _item = item;
-            _unsubscribe = unsubscribe;
-        }
-
-        public void Dispose()
-        {
-            _unsubscribe?.Invoke(_item);
-        }
+        _unsubscribe?.Invoke(_item);
     }
 }
