@@ -1,26 +1,19 @@
-﻿using Microsoft.Xna.Framework;
+﻿namespace MonoLibrary.Engine.Services.Updates;
 
-using MonoLibrary.Engine.Objects;
+//public interface IRendererService : IDisposable
+//{
+//    /// <summary>
+//    /// Ran after <see cref="Game.BeginDraw"/>.
+//    /// </summary>
+//    void BeforeDraw() { }
 
-using System;
+//    /// <summary>
+//    /// Ran before <see cref="Game.Draw(GameTime)"/>, hence before <see cref="GameObject.Draw(GameTime)"/>.
+//    /// </summary>
+//    void Draw(float deltaTime);
 
-namespace MonoLibrary.Engine.Services.Updates
-{
-    //public interface IRendererService : IDisposable
-    //{
-    //    /// <summary>
-    //    /// Ran after <see cref="Game.BeginDraw"/>.
-    //    /// </summary>
-    //    void BeforeDraw() { }
-
-    //    /// <summary>
-    //    /// Ran before <see cref="Game.Draw(GameTime)"/>, hence before <see cref="GameObject.Draw(GameTime)"/>.
-    //    /// </summary>
-    //    void Draw(float deltaTime);
-
-    //    /// <summary>
-    //    /// Ran before <see cref="Game.EndDraw"/>.
-    //    /// </summary>
-    //    void AfterDraw() { }
-    //}
-}
+//    /// <summary>
+//    /// Ran before <see cref="Game.EndDraw"/>.
+//    /// </summary>
+//    void AfterDraw() { }
+//}
